+++ AlchemistCode/Cards/Common/Congeal.cs
@@ -14,7 +14,7 @@ public class Congeal : AlchemistCard
     {
         WithCalculatedBlock(6, 1, static (card, _) =>
             card.Owner.Creature.GetPowerAmount<PoisonPower>(), ValueProp.Move, 3, 0);
-        WithPower<RegenPower>(2, 1);
+        WithPower<RegenPower>(3, 1);
         WithTip(typeof(PoisonPower));
         WithTip(typeof(RegenPower));
     }